Removed unnecessary file and renamed another
