+++ utils/nameAndTextTools.ts
